feat: Update DetailPage to use useParams and usePhoto hooks
This commit updates the DetailPage component to utilize the useParams and usePhoto hooks. The useParams hook is used to retrieve the "id" parameter from the URL, which is then passed to the usePhoto hook to fetch the corresponding photo data. This improves the functionality of the DetailPage by dynamically loading the selected photo based on the URL parameter.